Adding feature file for Order placement steps.
