+++ src/main/java/com/automation/pages/AddressPage.java
@@ -1,0 +1,4 @@
+package com.automation.pages;
+
+public class AddressPage {
+}
