+++ src/main/java/com/automation/pages/ShippingPage.java
@@ -1,0 +1,5 @@
+package com.automation.pages;
+
+public class ShippingPage {
+
+}
