+++ src/main/java/com/automation/steps/AddressConfirmationSteps.java
@@ -1,0 +1,15 @@
+package com.automation.steps;
+
+import cucumber.api.java.en.Then;
+import cucumber.api.java.en.When;
+
+public class AddressConfirmationSteps {
+    @Then("^Verify \"([^\"]*)\" address on Address confirmation page$")
+    public void verify_address_on_Address_confirmation_page(String arg1){
+
+    }
+    @Then("^User clicks on proceed to checkout button on Address confirmation page$")
+    public void user_clicks_on_proceed_to_checkout_button_on_Address_confirmation_page(){
+
+    }
+}
