+++ src/main/java/com/automation/steps/OrderConfirmationPageSteps.java
@@ -5,8 +5,14 @@ import cucumber.api.java.en.Then;
 
 public class OrderConfirmationPageSteps {
     OrderPage orderPage = new OrderPage();
-    @Then("^Verify order confirmation details on order confirmation page$")
+    @Then("^Verify order details on order confirmation page$")
     public void verifyOrderConfirmationDetails(){
 
     }
+/*    @Then("^Verify order details on order confirmation page$")
+    public void verify_order_details_on_order_confirmation_page(){
+
+    }*/
+
+
 }
