+++ src/main/java/com/automation/steps/OrderPageSteps.java
@@ -8,9 +8,5 @@ import cucumber.api.java.en.When;
 public class OrderPageSteps {
     OrderPage orderpage = new OrderPage();
 
-    @When("^User verifies product details on Order page$")
-    public void user_verifies_product_details_on_Order_page(){
-        orderpage.verifyOrderedProductDetail();
-    }
 
 }
