+++ src/main/java/com/automation/steps/ShippingPageSteps.java
@@ -1,0 +1,19 @@
+package com.automation.steps;
+
+import cucumber.api.java.en.Then;
+
+public class ShippingPageSteps {
+    @Then("^Verify terms of service on Shipping page$")
+    public void verify_terms_of_service_on_Shipping_page(){
+
+
+    }
+    @Then("^User click on agree button on Shipping page$")
+    public void user_click_on_agree_button_on_Shipping_page(){
+
+    }
+    @Then("^User click on proceed to checkout on Shipping page$")
+    public void user_click_on_proceed_to_checkout_on_Shipping_page(){
+
+    }
+}
