+++ src/test/java/features/placeanorder.feature
@@ -1,6 +1,7 @@
 Feature: Place an order
+
   @smk
-  Scenario: Verify successfull placement of any order
+  Scenario: Verify successful placement of any order
 
     Given User is on Homepage of the application
     When User navigates to login page
@@ -12,7 +13,22 @@ Feature: Place an order
     When User adds item to the cart from product detail page
     Then Verify "Printed Summer Dress" product on cart page
     When User proceeds to checkout from Cart page
-    When User verifies product details on Order page
-    Then Verify order confirmation details on order confirmation page
+
+    Then User verifies product details on Summary page
+    When User clicks on proceed to checkout button on Summary page
+
+    Then Verify "Delivery and Billing" address on Address confirmation page
+    Then User clicks on proceed to checkout button on Address confirmation page
+
+    Then Verify terms of service on Shipping page
+    And  User click on agree button on Shipping page
+    And  User click on proceed to checkout on Shipping page
+
+    And  Verify product details on Payment page
+    When select "Pay by check" as payment method on Payment page
+    Then Verify selected payment method on payment page
+
+    When User clicks on confirm my order button on payment page
+    Then Verify order details on order confirmation page
