+++ src/test/java/runner/TestRunner.java
@@ -7,7 +7,7 @@ import org.junit.runner.RunWith;
 @RunWith(Cucumber.class)
 @CucumberOptions(tags = "@smk", plugin = {"pretty", "html:target/cucumber-report",
         "json:target/cucumber.json"}, glue = "com.automation.steps",
-        dryRun = false,
+        dryRun = true,
         features = "src/test/java/features/")
 public class TestRunner {
